Move CustomDropdownView and CustomCheckboxView to more intuitive location
